# frontend/src/services/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NexusCAD Pro - Digital Twin API",
    "description": "REST API for Digital Twin / BIM / CAD application with device management, drawing synchronization, report generation, and CAD/BIM export capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "NexusCAD Pro Support",
      "email": "support@nexuscad.pro"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.nexuscad.pro/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List all projects",
        "operationId": "getProjects",
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/SortParam" },
          { "$ref": "#/components/parameters/OrderParam" }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponsePaginatedProjects" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "operationId": "createProject",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProject" }
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "operationId": "getProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdParam" }],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProject" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Projects"],
        "summary": "Update project",
        "operationId": "updateProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseProject" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete project",
        "operationId": "deleteProject",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdParam" }],
        "responses": {
          "204": { "description": "Project deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ProjectIdParam": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "default": 1, "minimum": 1 }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
      },
      "SortParam": {
        "name": "sort",
        "in": "query",
        "schema": { "type": "string", "default": "createdAt" }
      },
      "OrderParam": {
        "name": "order",
        "in": "query",
        "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
      "ApiResponseProject": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/Project" },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "ApiResponsePaginatedProjects": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/PaginatedProjects" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "PaginatedProjects": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "author": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["active", "archived", "draft"] },
          "deviceCount": { "type": "integer" },
          "connectionCount": { "type": "integer" }
        }
      },
      "CreateProjectInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "author": { "type": "string" }
        }
      },
      "UpdateProjectInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "archived", "draft"] }
        }
      },
...
    }
  }
}